+++ client/include/autd3/link/emulator.hpp
@@ -3,7 +3,7 @@
 // Created Date: 05/07/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 03/09/2021
+// Last Modified: 21/11/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -40,7 +40,7 @@ class Emulator : virtual public core::Link {
   void open() override = 0;
   void close() override = 0;
   void send(const uint8_t* buf, size_t size) override = 0;
-  void read(uint8_t* rx, size_t buffer_len) override = 0;
+  void receive(uint8_t* rx, size_t buffer_len) override = 0;
   bool is_open() override = 0;
 };
 }  // namespace autd::link